🏗️ Extract reactions list from `ReactionsCounter` component into new dedicated components
Having dedicated components, `ReactionList` and `ReactionsListItem`, is clearer than packing everything into `ReactionsCounter`. This approach better adheres to the separation of concerns principle.
