+++ src/features/activity-feed/reaction/components/ReactionsCounter/index.jsx
@@ -2,20 +2,9 @@ import { useState } from 'react'
 import PropTypes from 'prop-types'
 import './style.scss'
 import { useSelector } from 'react-redux'
-import { Box, Button, Popover, Typography, Link as MuiLink } from '@mui/material'
-import { Link } from 'react-router-dom'
-import { styled } from '@mui/material/styles'
-import Badge from '@mui/material/Badge'
-import Avatar from '@mui/material/Avatar'
-import { getUserOrganizationId } from '../../../../user/store/userSelectors'
+import { Button, Popover } from '@mui/material'
 import { getPostReactions } from '../../../../../redux/selectors/feed'
-
-const SmallAvatar = styled(Avatar)(({ theme }) => ({
-    width: 20,
-    height: 20,
-    background: `${theme.palette.background.paper}`,
-    padding: 2
-}))
+import { ReactionsList } from '../ReactionsList'
 
 ReactionsCounter.propTypes = {
     postId: PropTypes.number.isRequired
@@ -24,7 +13,7 @@ ReactionsCounter.propTypes = {
 export function ReactionsCounter({ postId }) {
     const [anchorEl, setAnchorEl] = useState(null)
     const postReactions = useSelector(getPostReactions(postId))
-    const organizationId = useSelector(getUserOrganizationId)
+
     const handleClick = event => {
         setAnchorEl(event.currentTarget)
     }
@@ -56,50 +45,7 @@ export function ReactionsCounter({ postId }) {
                     horizontal: 'left'
                 }}
             >
-                <Box className="c-reaction-post__info">
-                    {postReactions.map(reaction => (
-                        <Box
-                            className="c-reaction-post__info-emoji"
-                            key={reaction.id}
-                            sx={{ display: 'flex', alignItems: 'center', padding: '0', margin: 1 }}>
-                            <Badge
-                                className="c-reaction-post__info-container-picture"
-                                sx={{ marginRight: '0.5em' }}
-                                overlap="circular"
-                                anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-                                badgeContent={
-                                    <SmallAvatar src={`/assets/reactions/emoji-${reaction.type}.png`} />
-                                }
-                            >
-                                <Avatar
-                                    component={Link}
-                                    to={`/${organizationId}/user/${reaction.author.id}`}
-                                    alt={reaction.author.name}
-                                    src={reaction.author.profilePicture}
-                                />
-                            </Badge>
-                            <Box>
-                                <MuiLink
-                                    component={Link}
-                                    to={`/${organizationId}/user/${reaction.author.id}`}
-                                >
-                                    <Typography
-                                        variant="body2"
-                                        className="c-reaction-post__identity"
-                                    >
-                                        {`${reaction.author.name} ${reaction.author.surname}`}
-                                    </Typography>
-                                </MuiLink>
-                                <Typography
-                                    variant="body2"
-                                    className="c-reaction-post__job"
-                                >
-                                    {reaction.author.job}
-                                </Typography>
-                            </Box>
-                        </Box>
-                    ))}
-                </Box>
+                <ReactionsList postId={postId} />
             </Popover>
         </>
     )
